move bugs fetch into build_package, drop make_release_notes
